refactor: extract shared SessionInfo type to extensions/shared/types.ts
Both pidash-ui and pidash-server now import from the shared location.
Eliminates the duplicated 17-line interface.

Completes Fix 6 from issue #442 (previously deferred).
